requirements.txt
fastapi
uvicorn 
sqlalchemy
fastapi-sqlalchemy
python-dotenv
psycopg2
alembic
pydantic